Moved the logic for generating a homepage, into a separate file and directory
I did this because I will be reusing that logic in a future commmit.
However, this led to some unexpected CORS-related issues.
I therefore installed an extension in my development environment that lets me connect to my local server.
I also updated the README with usage instructions that include this information.