Created functions for submitting the signup form and for generating error lists
The signup function is invoked when the user submits the signup form.
Currently, it checks for a blank username and/or password and re-renders the signup page if there are errors.
If there are errors, the generateSignupPage function now calls a helper function generateErrorList to display these errors as a list.