Displayed HTTP 409 errors under the username input field
Until now, these errors were displayed at the end of the form.
This resulted in the same error message being added to the form each time a user tried to sign up with an existing username.
Additionally, older error messages related to usernames and passwords were not erased.
This commit fixes these issues by adding the HTTP 409 error to the list of username-related error messages and regenerating the signup form.
As noted in the comment, this code will be updated if there are other HTTP 409 error messages later on.