Called generateErrorList from the functions for generating signup and login pages instead of from createInputGroup
In my functions for generating login and registration pages, I had been passing a list of errors to createInputGroup.
CreateInputGroup then passed that same list of errors to generateErrorList.
By then, it was nearly impossible to tell what data type the list of errors was supposed to be.
This commit fixes that issue by calling generateErrorList from the functions for generating signup and login pages instead.