Minor edit: Changed the position of a button and changed the colors of certain buttons and messages
I wanted to put the logout button above the 'Brew Coffee With a Teapot' button in order to make its position consistent with the signup/login buttons.
Also, I made the signup, login, and logout buttons the same shade of blue to match the teacup image and the buttons on the forms.
Lastly, I made the success messages blue in order to match the teacup image; a darker shade was needed to make the text readable.