Created a function that fetches a concoction and displays it, and created corresponding buttons on the list of concoctions
Note: I did not yet include the code for handling HTTP 400 and 401 responses; this will be added in the next commit.
The "corresponding buttons" are the newly added "View Concoction" buttons next to each concoction on the "Your Concoctions" page; these buttons call this function.
More concoction-related information will be displayed on the concoction page in future commits.