Refactored the authentication and concoction files by creating a utility function that creates and appends success headings
This required me to rename the appendErrorHeading file to messageHeadings.
The idea is that this file can now be used to create different types of headings, not just error headings.
I also made the appendErrorHeading function a regular export instead of a default one, to reflect how this file should be used.
Finally, I included a small change to the generateLoginPage function to make the code a bit more concise.