Added username and validation bug fixes and removed the password from the input field for security purposes
This probably should have been two commits, but they are related.
When the username and/or password is invalid, I have now chosen not to print the password on the signup page for security purposes; this change will also be made to the login form in a later commit.
Because of how JavaScript defines truthy and falsy values, the signup form was displaying a "password must not match the username" error when the username was blank; this commit fixes that.
Additionally, the signup form now displays a "username/password is required" error message when the username or password contains only whitespace.
Finally, because I had forgotten to add a getter method for the username, it was not displaying on the signup form when there was an error; this commit fixes that bug.