Refactored the generateLoginPage function with the User class and removed the password from its input field when there is an error
Refactoring the generateLoginPage function also required me to update some of the functions that call it.
I decided to prevent the password from displaying in the password input field whenever there is a validation error.
I did this for security purposes, even though it's not a major issue (the password is not displayed in plain text).