Used the Concoction class to refactor the generateConcoctionPage function
This required me to add a few new Concoction methods and make use of the listItemId method.
I also changed a variable name to make it clear that I am fetching additional concoction data, not ALL of a concoction's data.
Note: This should (theoretically and eventually) make it easier to add coffee and ingredient data and generate an 'Edit Concoction' form.