Added and displayed each concoction's ingredients, sorted by category
This included creating an Ingredient class and enabling the Concoction instances to keep track of their ingredients and ingredient categories.
I also had to style each list of ingredients in order to keep the format consistent with coffee, instructions, etc.
I also slightly changed the coffee description in order to keep it consistent with the ingredient descriptions.