Fixed a bug with the addLiquid function and created a new button that removes liquid ingredients from the form
DO NOT REVERT PAST THIS COMMIT!!!
In the previous commit, I mistakenly used a global variable to track the number of liquid ingredient fields on the form.
The value of that variable increased instead of resetting to 0 every time the 'New Concoction' button was clicked.
This messed up the id values for the liquid fields, and it caused the newly-created 'Remove Liquid' button to display even when there was only one liquid ingredient.
(A concoction needs at least one liquid ingredient, so this button should only be displayed when there are two or more liquids.)
This commit fixes that issue by having the addLiquid function instead count the number of child elements in the list of liquids.