Created a heading and buttons for adding and removing additional ingredients in the New Concoction form
This required me to update a few other functions in order to account for the 'Additional Ingredient' category.
These updates included creating a regular divider instead of an ingredient divider if the ingredient category was 'Additional Ingredient'.
I also renamed a divider variable to keep it consistent with other similar variables.
Finally, I added a style to "remove ingredient"-related error messages in order to keep them centered in the form.