Simplified the concoction form functions by moving and combining similar code
The biggest change here is that I combined the code from the addLiquid, addSweetener, addCreamer, and addAdditionalIngredient functions,
and moved it into the addIngredient function; I was then able to delete the other functions.
I also made some functions use the ingredient category to determine the minIngredientCount and placeholder texts, instead of accepting them as arguments.
In the process, I was able to shorten a few function signatures and make the code much more readable.