Fixed a bug with a hidden input's value and simplified repetitive code
DO NOT REVERT PAST THIS COMMIT!!!
When the category sent to addIngredient is 'ingredient', the corresponding category input field's value should be 'Additional Ingredient'.
Instead, its value was 'Ingredient' because the function was mistakenly looking for a category whose value was 'additional'.
This commit fixed that issue and simplified some repetitive code.
Another utility file was created in the process: wordFunctions.js; this contains functions for capitalizing and lowercasing words.