Part 1 of the createConcoction function: Fixed a bug, prevented a bug, and converted form data into JSON
DO NOT REVERT BACK TO THIS COMMIT!!!
Until now, when a user deleted all of a category's ingredients from the form, the corresponding list was still set to 'display: block'.
This created an issue when trying to filter for lists containing ingredients.
This commit fixes that bug by resetting an empty list of ingredients to 'display: none' (its initial value) instead.

Also, all of the New Concoction form's ordered lists are currently lists of ingredients.
I gave these lists a class of their own, in order to avoid any bugs in the future
(i.e. in case I ever need to create and select any non-ingredient ordered lists on the form).