Part 2 of the createConcoction function: Validated concoction-specific attributes and added/removed corresponding error messages
I updated the Concoction class to check for an empty name and instructions. I then created code for adding error messages to the form.
I also added code that removes error messages when there are no longer any errors associated with a name and/or instructions.
Note that I also had to update other functions with code specifically related to a concoction's instructions.
This added a lot of repetitive code, so the next commit will simplify it.