Part 6 of createConcoction: Combined the logic for displaying and removing error messages for concoction and coffee inputs
This required changing a class on an <li> to an id instead, as well as changing the name of one of the Concoction error messages.
Because the code uses the same functions for adding/removing error messages in the same way for concoctions and coffees,
I was able to put similar input objects in the same array and iterate over it; this greatly simplified my code.
I also rewrote a for loop that had produced tight coupling by using the same index position to iterate over two arrays.
The new for loop now iterates over one array of coffee input names - much more robust and one less variable to worry about!