Part 8 of createConcoction: Added ingredient-related error messages to the New Concoction form
This required me to set an id on each list item containing an ingredient's category, amount, and name.
That then allowed me to simplify the code for getting the position of each ingredient in the 'ingredients' variable in createConcoction.
This also required me to update how I set each ingredient error's category in the Ingredient.validateIngredients() function.