Part 9 of createConcoction: Added and removed ingredient error messages depending on the state of the form
Note: This does NOT remove error headings when an ingredient has no errors at all; this will be implemented in the next commit.
It does, however, update the form if, e.g., an ingredient previously had a name and no amount, but now has an amount and no name.
It also adds an error heading if an ingredient previously had no errors but now has one.
There is a lot of repetitive code in this commit; this will be simplified later.