Part 10 of createConcoction: Restructured the list of ingredient errors and added or removed error messages as needed
I finally figured out that the best way to add and remove ingredient error messages, was to iterate over each ingredient list item.
This required me to change the Ingredient class by having it return an error object, where each key is an ingredient list item's id,
and its value is the corresponding error message.
Previously, it used a helper method and returned lists of ingredient error objects, each list corresponding to a category.
This restructuring enabled me to use a single for loop to iterate over each ingredient and update it based on the presence or lack of errors.
The next commit will simplify the code and account for the case when none of the ingredients has an error.