Fixed some bugs with an edge case related to buttons for adding and removing ingredients
In a previous commit, I tried to add an error message for the edge case when someone sets a 'Remove <Ingredient Category>' button from 'display: none' to 'display: initial'.
This did not entirely work, as I discovered later. The code added the error message to the end of a list of ingredients.
This directly impacted the calculation of the number of ingredients in a list of ingredients and introduced a few bugs.
In some cases, e.g., the error message didn't display, and the 'Add <Ingredient Category>' button didn't work.

I fixed these bugs by checking for an error heading at the end of a list of ingredients, before adding or removing an ingredient.
I also removed the error message when the user clicks on the button to add an ingredient.
Finally, I made the list of ingredients visible when an error message is created, which makes the error message visible as well.