Part 15 of createConcoction: Added code for handling HTTP 409 responses, primarily ones for duplicate concoction names
The only expected HTTP 409 response is when the user has created a concoction with a certain name and tries to create another with the same name.
Other HTTP 409 responses are displayed in the console.

I also removed a couple of TODOs, including one regarding the unlikely edge case where blank required values are sent to the backend.
This is unlikely because there are measures on the frontend to prevent this, and it would be a huge hassle to identify the blank fields and display error messages for them.
Those rare cases will instead be displayed in the console.