Part 16 of createConcoction: Updated hasErrorHeading and displayed error messages for an HTTP 400 edge case
Please ignore part of the description in the previous commit.
I decided to display a list of error messages in the unlikely edge case that an HTTP 400 response will be returned for missing data.
This also required me to update hasErrorHeading (and some areas of code that use it), which then made the function's purpose clearer and more accurate.