Moved code for displaying a concoction, into its own function, and removed coffee-related data from the Concoction class
My goal is for this new function to be used by the generateConcoctionPage and createConcoction functions.
The createConcoction function returns concoction, coffee, and ingredient data separately, so it makes more sense to keep this data separate.
Thus, I have removed coffee-related data from the Concoction class, and I also plan to remove ingredient-related data from it.