Moved the list of ingredient categories from the Concoction class to the Ingredient class
The responsibility of keeping track of each ingredient category, belongs to the Ingredient class, not the Concoction class.
As for my idea in the previous commit, it didn't entirely work (aside from moving the list of ingredients to the Ingredient class).
I tried to make a static method Ingredient.ingredients(), which (in my mind) would work similarly to Concoction's 'get ingredients()' method.
However, because the Ingredient class didn't keep track of the concoction id, Ingredient.ingredients() returned EVERY concoction's ingredients!
I have decided to instead let every instance of the Concoction class keep track of its ingredients; this is much simpler.