Prevented the login function from re-creating the login page when the username and/or password is missing
DO NOT REVERT TO THIS COMMIT! Two different error lists are created here, which may lead to some bugs with displaying/removing error messages.
When I created the concoctions, I realized that I didn't need to re-create the whole concoction form when there was an error.
The same can be said of the user signup and login functions. Therefore, I will be refactoring them in the following commits.