Finished preventing the login function from re-creating the login page, by updating it for other HTTP return statuses
I needed different code for HTTP 400 and 401 error statuses because I finally realized that HTTP 401 only gets returned for invalid passwords.
This commit also fixes the "bugs with displaying/removing error messages" mentioned in the previous commit.

I also moved the line of code that resets the password to a blank value when there is an error.
Formerly, it would only do this if the password was blank.