Combined two functions, updated a function, and made login faster by moving the code that removes error lists
I effectively replaced handleMissingInput with removeErrorList because they worked almost exactly the same way.
The only difference was that handleMissingInput had a different error message and checked for empty user input.
I also renamed removeErrorList to removeErrorLists and updated its parameters so that it could execute the same code on multiple error lists.

I moved the code that calls removeErrorList/s because I didn't want it to execute when the server returned an HTTP 200 or 500 status.
That wouldn't be necessary, since a new page would be created; in the case of HTTP 200, it would also delay user login.