Removed userInfo as a parameter from generateSignupPage, generateLoginPage, and createUserInputGroup
Originally, when there were username and or/password errors, the signup and login functions would recreate the signup and login pages.
They would do this by sending the corresponding 'generate' functions an argument named 'userInfo' that contained error messages.
The userInfo argument would then be sent to the createUserInputGroup function, which would use userInfo to create and display a list of errors.
Since the signup and login functions now update the username and password parts of a form, they don't call generateSignupPage or generateLoginPage.
Thus, the userInfo argument/parameter is no longer necessary.