Prepended certain error headings to the tops of their wrappers in order to make them more visible to the user
This required creating a prependErrorHeading function and having certain functions call it instead of appendErrorHeading.
The error heading created by prependErrorHeading also has some unique CSS styles, so I added a CSS class for it.
Finally, I created a helper method createErrorHeadingForWrapper and used it to refactor prependErrorHeading and appendErrorHeading.