Moved code for displaying concoctions into its own function "displayConcoctions"
I did for generateConcoctionsPage what I did for generateConcoctionPage: separate the code that fetches concoction data, from the code that displays it.
However, I still kept the code for generating the page title and displaying the login success message, in the generateConcoctionsPage function.
This is because I still want to display the "Your Concoctions" page even if there is an error while fetching concoctions.